Updated the cdk app by creating 3 independent app. Resolve review comments
