+++ java/codebuild/reportgroup/README.md
@@ -1,0 +1,18 @@
+# Welcome to your CDK Java project!
+
+It is a Maven-based project, so you can open this directory with any Maven-compatible Java IDE, and you should be able to build and run tests from your IDE.
+
+You should explore the contents of this template. It demonstrates a CDK app to create a codebuild ReportGroup cloud formation resource.
+
+The cdk.json file tells the CDK Toolkit how to execute your app. This example relies on maven to do that.
+
+## Useful commands
+
+ * `mvn package`     compile and run tests
+ * `cdk ls`          list all stacks in the app
+ * `cdk synth`       emits the synthesized CloudFormation template
+ * `cdk deploy`      deploy this stack to your default AWS account/region
+ * `cdk diff`        compare deployed stack with current state
+ * `cdk docs`        open CDK documentation
+
+Enjoy!
