+++ java/codebuild/sourceCredential/cdk.json
@@ -1,0 +1,3 @@
+{
+  "app": "mvn -e -q exec:java"
+}
